Adds base for a new config command
This is still not complete. I need to now actually make the set command
write arguments to relavent config files.

I also learned that I wasn't handling validation for the "backend"
field. I added a "quick fix" in the `latz.cli` module that I plan on
going back to later to refactor so that it's organized better.